requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi
slowapi
uvicorn
pydantic
requests
torchvision
Pillow
python-multipart
pytest
httpx
pytest-asyncio